extracted training pipeline
